# reduce NumPy overhead in ScaleTranslate and Affine transform calls